Refactory project structure
