+++ server/app/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from app.main.service import userService
+
+app.run(debug=True)
